Remove extra files and code
