+++ app/Http/Controllers/Controller.php
@@ -6,8 +6,29 @@ use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\MessageBag;
 
 abstract class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+    /**
+     * Message bag.
+     *
+     * @var Illuminate\Support\MessageBag
+     */
+    protected $messageBag = null;
+
+    /**
+     * Initializer.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        // CSRF Protection
+        $this->beforeFilter('csrf', array('on' => 'post'));
+
+        //
+        $this->messageBag = new MessageBag;
+    }
 }
